+++ package.json
@@ -37,8 +37,6 @@
     "start": "node ./src/bot.js",
     "fill": "node ./scripts/filldb.js",
     "init": "node ./scripts/initdb.js",
-    "fillv2": "node ./scripts/filldb2.js",
-    "initv2": "node ./scripts/initdb2.js",
     "web": "node ./src/web/Server.js",
     "gen-legacy-chat": "node ./scripts/legacy/discord.js",
     "gen-legacy-game": "node ./scripts/legacy/game.js",
